Fix import in cli and move qualifier and delivery logic into correct namespaces
