+++ src/backend/Dockerfile.cli
@@ -36,6 +36,8 @@ WORKDIR /srv/caliopen/src/backend/main/py.main
 RUN python setup.py develop
 WORKDIR /srv/caliopen/src/backend/components/py.pi
 RUN python setup.py develop
+WORKDIR /srv/caliopen/src/backend/interfaces/NATS/py.client
+RUN python setup.py develop
 
 ## Container specific installation part
 WORKDIR /srv/caliopen/src/backend/tools/py.CLI
