+++ src/backend/tools/py.CLI/setup.py
@@ -13,8 +13,8 @@ with open(os.path.join(*([here] + name.split('.') + ['__init__.py']))) as v_file
     version = re.compile(r".*__version__ = '(.*?)'", re.S).match(v_file.read()).group(1)
 
 requires = [
-    'caliopen_main',
-    'caliopen_storage',
+    'caliopen_pi',
+    'caliopen_nats',
 ]
 
 setup(name=name,
